Add a sync byte to each command
Each command is now 3 instead of 2 bytes. A third byte, the "sync byte",
is added to the end of the message. It has the value 0xFF and is used to
verify that all bytes were sent properly. This fixes the problem that
if one byte was missed, all subsequent messages would be messed up due
to the first message not having enough bytes, and thus remaining in the
buffer.

+++ src/uart.c
@@ -19,7 +19,7 @@ void UART_InterruptInit() {
 	ES = 1;
 }
 
-unsigned short UART_Buffer = 0;
+unsigned long UART_Buffer = 0;
 void UART_InterruptRoutine() interrupt 4 using 1 {
 	//Clear receive flag
 	RI = 0;
